lessons-3.1

+++ index.php
@@ -79,7 +79,9 @@ $products = mysqli_fetch_all($result, MYSQLI_ASSOC);
                         <h2 class="card-title"><?= $product ['name'] ?></h2>
                         <p class="card-text"><?= $product ['description'] ?></p>
                         <a href="/product.php?id=<?= $product['id'] ?>">Подробнее</a>
-                        <a class="btn btn-danger" href="/delete.php?id=<?= $product['id'] ?>">Удалить</a>
+                        <?php if (isLogin()): ?>
+                            <a class="btn btn-danger" href="/delete.php?id=<?= $product['id'] ?>">Удалить</a>
+                        <?php endif ?>
                     </div>
                 </div>
             </div>
